+++ docs/en_US/changelog.md
@@ -4,6 +4,12 @@
 >
 >As a reminder if there is no information on the update, it means that it only concerns the updating of documentation, translation or text.
 
+# 15/01/2021
+- Regular expression fix for comparison data
+- New widget's template (Jazpar 4)
+- New action to refresh data by querying the website (outside of the cron)
+- Option to use dates in widgets instead of command names
+
 # 08/01/2021
 - New widget's template
 - Better i18n