+++ docs/fr_FR/changelog.md
@@ -4,6 +4,12 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 15/01/2021
+- Correction de l'expression régulière pour les données de comparaison
+- Nouveau template de widget (Jazpar 4)
+- Nouvelle action pour rafraichir les données en interrogeant le site (hors cron)
+- Nouvelle option pour utiliser des dates plutôt que les noms des commandes dans les widgets
+
 # 08/01/2021
 - Nouveau template de widget 
 - Meilleure i18n